refactor DownloadLatestTranslation composable
